requirements.txt
# Core
anthropic>=0.40.0
python-dotenv>=1.0.0

# CLI
rich>=13.0.0
prompt_toolkit>=3.0.0

# Web UI
fastapi>=0.115.0
uvicorn>=0.32.0

# Testing
pytest>=8.0.0
